# shiftory-server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Shiftory API
  version: 1.0.0
  description: 排班协同、导入预览与团队日历 API。业务接口统一返回 `{data, requestId}`，错误返回 `{error, requestId}`。
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /workspaces/{workspaceId}/calendar:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [Calendar]
      operationId: getCalendar
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - name: memberIds
          in: query
          schema: { type: string, example: '1,2,3' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
    StartDate:
      name: start
      in: query
      required: true
      schema: { type: string, format: date }
    EndDate:
      name: end
      in: query
      required: true
      schema: { type: string, format: date }
...
  responses:
    Success:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [data, requestId]
      properties:
        data: {}
        requestId: { type: string, format: uuid }
...